stock_app/requirements.txt
pandas
numpy
matplotlib
requests
ta
python-dotenv
scipy
python-dotenv
alpha_vantage
statsmodels
mplfinance
yfinance
seaborn
